Finish the bot for RoomPlaza
The function takes a city as a parameter, then checks all the listings from RoomPlaza from that city. After that it compares their ids to the ids from the last time it checked (they are stored int a txt file) and returns the ids and links to the pages of the listings which do not have their ids in the txt file. After that, the txt file is updated with the new ids.